+++ .eslintrc.js
@@ -95,7 +95,6 @@ module.exports = {
       files: [
         'packages/jest-jasmine2/src/jasmine/**/*',
         'packages/expect/src/jasmineUtils.ts',
-        'e2e/browser-support/browserTest.js',
         '**/vendor/**/*',
       ],
       rules: {
@@ -142,6 +141,13 @@ module.exports = {
     // https://github.com/benmosher/eslint-plugin-import/issues/645
     'import/order': 0,
     'no-console': 0,
+    'no-restricted-imports': [
+      2,
+      {
+        message: 'Please use graceful-fs instead.',
+        name: 'fs',
+      },
+    ],
     'no-unused-vars': 2,
     'prettier/prettier': 2,
     'sort-imports': [2, {ignoreDeclarationSort: true}],