+++ packages/jest-resolve/src/defaultResolver.ts
@@ -5,11 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import * as fs from 'fs';
+import * as fs from 'graceful-fs';
 import {sync as resolveSync} from 'resolve';
-import {sync as browserResolve} from 'browser-resolve';
-import {sync as realpath} from 'realpath-native';
 import pnpResolver from 'jest-pnp-resolver';
+import {tryRealpath} from 'jest-util';
 import type {Config} from '@jest/types';
 
 type ResolverOptions = {
@@ -31,9 +30,7 @@ export default function defaultResolver(
     return pnpResolver(path, options);
   }
 
-  const resolve = options.browser ? browserResolve : resolveSync;
-
-  const result = resolve(path, {
+  const result = resolveSync(path, {
     basedir: options.basedir,
     extensions: options.extensions,
     isDirectory,
@@ -98,17 +95,7 @@ function realpathCached(path: Config.Path): Config.Path {
     return result;
   }
 
-  try {
-    result = realpath(path);
-  } catch (error) {
-    if (error.code !== 'ENOENT') {
-      throw error;
-    }
-  }
-
-  if (!result) {
-    result = path;
-  }
+  result = tryRealpath(path);
 
   checkedRealpathPaths.set(path, result);