+++ packages/jest-runtime/src/index.ts
@@ -23,7 +23,6 @@ import type {Config, Global} from '@jest/types';
 import type {
   Jest,
   JestEnvironment,
-  LocalModuleRequire,
   Module,
   ModuleWrapper,
 } from '@jest/environment';
@@ -54,8 +53,8 @@ import Snapshot = require('jest-snapshot');
 import stripBOM = require('strip-bom');
 
 interface JestGlobalsValues extends Global.TestFrameworkGlobals {
-  jest: JestGlobals.jest;
-  expect: JestGlobals.expect;
+  jest: typeof JestGlobals.jest;
+  expect: typeof JestGlobals.expect;
 }
 
 type HasteMapOptions = {
@@ -83,10 +82,17 @@ type InitialModule = Partial<Module> &
 type ModuleRegistry = Map<string, InitialModule | Module>;
 type ResolveOptions = Parameters<typeof require.resolve>[1];
 
-type BooleanObject = Record<string, boolean>;
-type CacheFS = {[path: string]: string};
+type StringMap = Map<string, string>;
+type BooleanMap = Map<string, boolean>;
 
-type RequireCache = {[key: string]: Module};
+const fromEntries: typeof Object.fromEntries =
+  Object.fromEntries ??
+  function fromEntries<T>(iterable: Iterable<[string, T]>) {
+    return [...iterable].reduce<Record<string, T>>((obj, [key, val]) => {
+      obj[key] = val;
+      return obj;
+    }, {});
+  };
 
 namespace Runtime {
   export type Context = JestContext;
@@ -122,18 +128,19 @@ const runtimeSupportsVmModules = typeof SyntheticModule === 'function';
 
 /* eslint-disable-next-line no-redeclare */
 class Runtime {
-  private _cacheFS: CacheFS;
+  private _cacheFS: StringMap;
   private _config: Config.ProjectConfig;
   private _coverageOptions: ShouldInstrumentOptions;
   private _currentlyExecutingModulePath: string;
   private _environment: JestEnvironment;
-  private _explicitShouldMock: BooleanObject;
+  private _explicitShouldMock: BooleanMap;
   private _internalModuleRegistry: ModuleRegistry;
   private _isCurrentlyExecutingManualMock: string | null;
-  private _mockFactories: Record<string, () => unknown>;
-  private _mockMetaDataCache: {
-    [key: string]: jestMock.MockFunctionMetadata<unknown, Array<unknown>>;
-  };
+  private _mockFactories: Map<string, () => unknown>;
+  private _mockMetaDataCache: Map<
+    string,
+    jestMock.MockFunctionMetadata<unknown, Array<unknown>>
+  >;
   private _mockRegistry: Map<string, any>;
   private _isolatedMockRegistry: Map<string, any> | null;
   private _moduleMocker: typeof jestMock;
@@ -142,28 +149,27 @@ class Runtime {
   private _esmoduleRegistry: Map<string, Promise<VMModule>>;
   private _resolver: Resolver;
   private _shouldAutoMock: boolean;
-  private _shouldMockModuleCache: BooleanObject;
-  private _shouldUnmockTransitiveDependenciesCache: BooleanObject;
-  private _sourceMapRegistry: SourceMapRegistry;
+  private _shouldMockModuleCache: BooleanMap;
+  private _shouldUnmockTransitiveDependenciesCache: BooleanMap;
+  private _sourceMapRegistry: StringMap;
   private _scriptTransformer: ScriptTransformer;
   private _fileTransforms: Map<string, TransformResult>;
   private _v8CoverageInstrumenter: CoverageInstrumenter | undefined;
   private _v8CoverageResult: V8Coverage | undefined;
-  private _transitiveShouldMock: BooleanObject;
+  private _transitiveShouldMock: BooleanMap;
   private _unmockList: RegExp | undefined;
-  private _virtualMocks: BooleanObject;
+  private _virtualMocks: BooleanMap;
   private _moduleImplementation?: typeof nativeModule.Module;
   private jestObjectCaches: Map<string, Jest>;
-  private _hasWarnedAboutRequireCacheModification = false;
 
   constructor(
     config: Config.ProjectConfig,
     environment: JestEnvironment,
     resolver: Resolver,
-    cacheFS?: CacheFS,
+    cacheFS: Record<string, string> = {},
     coverageOptions?: ShouldInstrumentOptions,
   ) {
-    this._cacheFS = cacheFS || Object.create(null);
+    this._cacheFS = new Map(Object.entries(cacheFS));
     this._config = config;
     this._coverageOptions = coverageOptions || {
       changedFiles: undefined,
@@ -171,13 +177,14 @@ class Runtime {
       collectCoverageFrom: [],
       collectCoverageOnlyFrom: undefined,
       coverageProvider: 'babel',
+      sourcesRelatedToTestsInChangedFiles: undefined,
     };
     this._currentlyExecutingModulePath = '';
     this._environment = environment;
-    this._explicitShouldMock = Object.create(null);
+    this._explicitShouldMock = new Map();
     this._internalModuleRegistry = new Map();
     this._isCurrentlyExecutingManualMock = null;
-    this._mockFactories = Object.create(null);
+    this._mockFactories = new Map();
     this._mockRegistry = new Map();
     // during setup, this cannot be null (and it's fine to explode if it is)
     this._moduleMocker = this._environment.moduleMocker!;
@@ -188,15 +195,15 @@ class Runtime {
     this._resolver = resolver;
     this._scriptTransformer = new ScriptTransformer(config);
     this._shouldAutoMock = config.automock;
-    this._sourceMapRegistry = Object.create(null);
+    this._sourceMapRegistry = new Map();
     this._fileTransforms = new Map();
-    this._virtualMocks = Object.create(null);
+    this._virtualMocks = new Map();
     this.jestObjectCaches = new Map();
 
-    this._mockMetaDataCache = Object.create(null);
-    this._shouldMockModuleCache = Object.create(null);
-    this._shouldUnmockTransitiveDependenciesCache = Object.create(null);
-    this._transitiveShouldMock = Object.create(null);
+    this._mockMetaDataCache = new Map();
+    this._shouldMockModuleCache = new Map();
+    this._shouldUnmockTransitiveDependenciesCache = new Map();
+    this._transitiveShouldMock = new Map();
 
     this._unmockList = unmockRegExpCache.get(config);
     if (!this._unmockList && config.unmockedModulePathPatterns) {
@@ -207,13 +214,11 @@ class Runtime {
     }
 
     if (config.automock) {
+      const virtualMocks = fromEntries(this._virtualMocks);
       config.setupFiles.forEach(filePath => {
         if (filePath && filePath.includes(NODE_MODULES)) {
-          const moduleID = this._resolver.getModuleID(
-            this._virtualMocks,
-            filePath,
-          );
-          this._transitiveShouldMock[moduleID] = false;
+          const moduleID = this._resolver.getModuleID(virtualMocks, filePath);
+          this._transitiveShouldMock.set(moduleID, false);
         }
       });
     }
@@ -280,7 +285,6 @@ class Runtime {
       mocksPattern: escapePathForRegex(path.sep + '__mocks__' + path.sep),
       name: config.name,
       platforms: config.haste.platforms || ['ios', 'android'],
-      providesModuleNodeModules: config.haste.providesModuleNodeModules,
       resetCache: options && options.resetCache,
       retainAllFiles: false,
       rootDir: config.rootDir,
@@ -296,7 +300,6 @@ class Runtime {
     moduleMap: HasteMap.ModuleMap,
   ): Resolver {
     return new Resolver(moduleMap, {
-      browser: config.browser,
       defaultPlatform: config.haste.defaultPlatform,
       extensions: config.moduleFileExtensions.map(extension => '.' + extension),
       hasCoreModules: true,
@@ -342,13 +345,13 @@ class Runtime {
         return core;
       }
 
-      const transformedFile = this.transformFile(modulePath, {
+      const transformedCode = this.transformFile(modulePath, {
         isInternalModule: false,
         supportsDynamicImport: true,
         supportsStaticESM: true,
       });
 
-      const module = new SourceTextModule(transformedFile.code, {
+      const module = new SourceTextModule(transformedCode, {
         context,
         identifier: modulePath,
         importModuleDynamically: this.linkModules.bind(this),
@@ -392,16 +395,15 @@ class Runtime {
       return globals;
     }
 
-    const resolved = this._resolveModule(
-      referencingModule.identifier,
-      specifier,
-    );
+    const [path, query] = specifier.split('?');
+
+    const resolved = this._resolveModule(referencingModule.identifier, path);
 
     if (
       this._resolver.isCoreModule(resolved) ||
       this.unstable_shouldLoadAsEsm(resolved)
     ) {
-      return this.loadEsmModule(resolved);
+      return this.loadEsmModule(resolved, query);
     }
 
     return this.loadCjsAsEsm(
@@ -420,9 +422,11 @@ class Runtime {
       'You need to run with a version of node that supports ES Modules in the VM API.',
     );
 
-    const modulePath = this._resolveModule(from, moduleName);
+    const [path, query] = (moduleName ?? '').split('?');
 
-    return this.loadEsmModule(modulePath);
+    const modulePath = this._resolveModule(from, path);
+
+    return this.loadEsmModule(modulePath, query);
   }
 
   private async loadCjsAsEsm(
@@ -458,7 +462,7 @@ class Runtime {
     isRequireActual?: boolean | null,
   ): T {
     const moduleID = this._resolver.getModuleID(
-      this._virtualMocks,
+      fromEntries(this._virtualMocks),
       from,
       moduleName,
     );
@@ -470,12 +474,12 @@ class Runtime {
     const manualMock =
       moduleName && this._resolver.getMockModule(from, moduleName);
     if (
-      (!options || !options.isInternalModule) &&
+      !options?.isInternalModule &&
       !isRequireActual &&
       !moduleResource &&
       manualMock &&
       manualMock !== this._isCurrentlyExecutingManualMock &&
-      this._explicitShouldMock[moduleID] !== false
+      this._explicitShouldMock.get(moduleID) !== false
     ) {
       modulePath = manualMock;
     }
@@ -490,7 +494,9 @@ class Runtime {
 
     let moduleRegistry;
 
-    if (!options || !options.isInternalModule) {
+    if (options?.isInternalModule) {
+      moduleRegistry = this._internalModuleRegistry;
+    } else {
       if (
         this._moduleRegistry.get(modulePath) ||
         !this._isolatedModuleRegistry
@@ -499,8 +505,6 @@ class Runtime {
       } else {
         moduleRegistry = this._isolatedModuleRegistry;
       }
-    } else {
-      moduleRegistry = this._internalModuleRegistry;
     }
 
     const module = moduleRegistry.get(modulePath);
@@ -546,7 +550,7 @@ class Runtime {
 
   requireMock<T = unknown>(from: Config.Path, moduleName: string): T {
     const moduleID = this._resolver.getModuleID(
-      this._virtualMocks,
+      fromEntries(this._virtualMocks),
       from,
       moduleName,
     );
@@ -562,8 +566,9 @@ class Runtime {
 
     const mockRegistry = this._isolatedMockRegistry || this._mockRegistry;
 
-    if (moduleID in this._mockFactories) {
-      const module = this._mockFactories[moduleID]();
+    if (this._mockFactories.has(moduleID)) {
+      // has check above makes this ok
+      const module = this._mockFactories.get(moduleID)!();
       mockRegistry.set(moduleID, module);
       return module as T;
     }
@@ -640,7 +645,7 @@ class Runtime {
     moduleRegistry: ModuleRegistry,
   ) {
     if (path.extname(modulePath) === '.json') {
-      const text = stripBOM(fs.readFileSync(modulePath, 'utf8'));
+      const text = stripBOM(this.readFile(modulePath));
 
       const transformedFile = this._scriptTransformer.transformJson(
         modulePath,
@@ -717,12 +722,17 @@ class Runtime {
     try {
       fn();
     } finally {
+      // might be cleared within the callback
+      this._isolatedModuleRegistry?.clear();
+      this._isolatedMockRegistry?.clear();
       this._isolatedModuleRegistry = null;
       this._isolatedMockRegistry = null;
     }
   }
 
   resetModules(): void {
+    this._isolatedModuleRegistry?.clear();
+    this._isolatedMockRegistry?.clear();
     this._isolatedModuleRegistry = null;
     this._isolatedMockRegistry = null;
     this._mockRegistry.clear();
@@ -798,7 +808,7 @@ class Runtime {
   }
 
   getSourceMaps(): SourceMapRegistry {
-    return this._sourceMapRegistry;
+    return fromEntries(this._sourceMapRegistry);
   }
 
   setMock(
@@ -807,17 +817,18 @@ class Runtime {
     mockFactory: () => unknown,
     options?: {virtual?: boolean},
   ): void {
-    if (options && options.virtual) {
+    if (options?.virtual) {
       const mockPath = this._resolver.getModulePath(from, moduleName);
-      this._virtualMocks[mockPath] = true;
+
+      this._virtualMocks.set(mockPath, true);
     }
     const moduleID = this._resolver.getModuleID(
-      this._virtualMocks,
+      fromEntries(this._virtualMocks),
       from,
       moduleName,
     );
-    this._explicitShouldMock[moduleID] = true;
-    this._mockFactories[moduleID] = mockFactory;
+    this._explicitShouldMock.set(moduleID, true);
+    this._mockFactories.set(moduleID, mockFactory);
   }
 
   restoreAllMocks(): void {
@@ -830,6 +841,32 @@ class Runtime {
 
   clearAllMocks(): void {
     this._moduleMocker.clearAllMocks();
+  }
+
+  teardown(): void {
+    this.restoreAllMocks();
+    this.resetAllMocks();
+    this.resetModules();
+
+    this._internalModuleRegistry.clear();
+    this._mockFactories.clear();
+    this._mockMetaDataCache.clear();
+    this._shouldMockModuleCache.clear();
+    this._shouldUnmockTransitiveDependenciesCache.clear();
+    this._explicitShouldMock.clear();
+    this._transitiveShouldMock.clear();
+    this._virtualMocks.clear();
+    this._cacheFS.clear();
+    this._unmockList = undefined;
+
+    this._sourceMapRegistry.clear();
+
+    this._fileTransforms.clear();
+    this.jestObjectCaches.clear();
+
+    this._v8CoverageResult = [];
+    this._v8CoverageInstrumenter = undefined;
+    this._moduleImplementation = undefined;
   }
 
   private _resolveModule(from: Config.Path, to?: string) {
@@ -936,7 +973,7 @@ class Runtime {
       value: this._createRequireImplementation(localModule, options),
     });
 
-    const transformedFile = this.transformFile(filename, options);
+    const transformedCode = this.transformFile(filename, options);
 
     let compiledFunction: ModuleWrapper | null = null;
 
@@ -945,38 +982,21 @@ class Runtime {
       const vmContext = this._environment.getVmContext();
 
       if (vmContext) {
-        if (typeof compileFunction === 'function') {
-          try {
-            compiledFunction = compileFunction(
-              transformedFile.code,
-              this.constructInjectedModuleParameters(),
-              {
-                filename,
-                parsingContext: vmContext,
-              },
-            ) as ModuleWrapper;
-          } catch (e) {
-            throw handlePotentialSyntaxError(e);
-          }
-        } else {
-          const script = this.createScriptFromCode(
-            transformedFile.code,
-            filename,
-          );
-
-          const runScript = script.runInContext(
-            vmContext,
-          ) as RunScriptEvalResult;
-
-          if (runScript === null) {
-            compiledFunction = null;
-          } else {
-            compiledFunction = runScript[EVAL_RESULT_VARIABLE];
-          }
+        try {
+          compiledFunction = compileFunction(
+            transformedCode,
+            this.constructInjectedModuleParameters(),
+            {
+              filename,
+              parsingContext: vmContext,
+            },
+          ) as ModuleWrapper;
+        } catch (e) {
+          throw handlePotentialSyntaxError(e);
         }
       }
     } else {
-      const script = this.createScriptFromCode(transformedFile.code, filename);
+      const script = this.createScriptFromCode(transformedCode, filename);
 
       const runScript = this._environment.runScript<RunScriptEvalResult>(
         script,
@@ -1029,22 +1049,28 @@ class Runtime {
     this._currentlyExecutingModulePath = lastExecutingModulePath;
   }
 
-  private transformFile(filename: string, options?: InternalModuleOptions) {
+  private transformFile(
+    filename: string,
+    options?: InternalModuleOptions,
+  ): string {
+    const source = this.readFile(filename);
+
+    if (options?.isInternalModule) {
+      return source;
+    }
+
     const transformedFile = this._scriptTransformer.transform(
       filename,
       this._getFullTransformationOptions(options),
-      this._cacheFS[filename],
+      source,
     );
 
-    // we only care about non-internal modules
-    if (!options || !options.isInternalModule) {
-      this._fileTransforms.set(filename, transformedFile);
-    }
+    this._fileTransforms.set(filename, transformedFile);
 
     if (transformedFile.sourceMapPath) {
-      this._sourceMapRegistry[filename] = transformedFile.sourceMapPath;
+      this._sourceMapRegistry.set(filename, transformedFile.sourceMapPath);
     }
-    return transformedFile;
+    return transformedFile.code;
   }
 
   private createScriptFromCode(scriptSource: string, filename: string) {
@@ -1166,12 +1192,14 @@ class Runtime {
     const modulePath =
       this._resolver.resolveStubModuleName(from, moduleName) ||
       this._resolveModule(from, moduleName);
-    if (!(modulePath in this._mockMetaDataCache)) {
+    if (!this._mockMetaDataCache.has(modulePath)) {
       // This allows us to handle circular dependencies while generating an
       // automock
 
-      this._mockMetaDataCache[modulePath] =
-        this._moduleMocker.getMetadata({}) || {};
+      this._mockMetaDataCache.set(
+        modulePath,
+        this._moduleMocker.getMetadata({}) || {},
+      );
 
       // In order to avoid it being possible for automocking to potentially
       // cause side-effects within the module environment, we need to execute
@@ -1195,36 +1223,39 @@ class Runtime {
             `See: https://jestjs.io/docs/manual-mocks.html#content`,
         );
       }
-      this._mockMetaDataCache[modulePath] = mockMetadata;
+      this._mockMetaDataCache.set(modulePath, mockMetadata);
     }
     return this._moduleMocker.generateFromMetadata(
-      this._mockMetaDataCache[modulePath],
+      // added above if missing
+      this._mockMetaDataCache.get(modulePath)!,
     );
   }
 
-  private _shouldMock(from: Config.Path, moduleName: string) {
+  private _shouldMock(from: Config.Path, moduleName: string): boolean {
     const explicitShouldMock = this._explicitShouldMock;
     const moduleID = this._resolver.getModuleID(
-      this._virtualMocks,
+      fromEntries(this._virtualMocks),
       from,
       moduleName,
     );
     const key = from + path.delimiter + moduleID;
 
-    if (moduleID in explicitShouldMock) {
-      return explicitShouldMock[moduleID];
+    if (explicitShouldMock.has(moduleID)) {
+      // guaranteed by `has` above
+      return explicitShouldMock.get(moduleID)!;
     }
 
     if (
       !this._shouldAutoMock ||
       this._resolver.isCoreModule(moduleName) ||
-      this._shouldUnmockTransitiveDependenciesCache[key]
+      this._shouldUnmockTransitiveDependenciesCache.get(key)
     ) {
       return false;
     }
 
-    if (moduleID in this._shouldMockModuleCache) {
-      return this._shouldMockModuleCache[moduleID];
+    if (this._shouldMockModuleCache.has(moduleID)) {
+      // guaranteed by `has` above
+      return this._shouldMockModuleCache.get(moduleID)!;
     }
 
     let modulePath;
@@ -1233,70 +1264,57 @@ class Runtime {
     } catch (e) {
       const manualMock = this._resolver.getMockModule(from, moduleName);
       if (manualMock) {
-        this._shouldMockModuleCache[moduleID] = true;
+        this._shouldMockModuleCache.set(moduleID, true);
         return true;
       }
       throw e;
     }
 
     if (this._unmockList && this._unmockList.test(modulePath)) {
-      this._shouldMockModuleCache[moduleID] = false;
+      this._shouldMockModuleCache.set(moduleID, false);
       return false;
     }
 
     // transitive unmocking for package managers that store flat packages (npm3)
     const currentModuleID = this._resolver.getModuleID(
-      this._virtualMocks,
+      fromEntries(this._virtualMocks),
       from,
     );
     if (
-      this._transitiveShouldMock[currentModuleID] === false ||
+      this._transitiveShouldMock.get(currentModuleID) === false ||
       (from.includes(NODE_MODULES) &&
         modulePath.includes(NODE_MODULES) &&
         ((this._unmockList && this._unmockList.test(from)) ||
-          explicitShouldMock[currentModuleID] === false))
+          explicitShouldMock.get(currentModuleID) === false))
     ) {
-      this._transitiveShouldMock[moduleID] = false;
-      this._shouldUnmockTransitiveDependenciesCache[key] = true;
+      this._transitiveShouldMock.set(moduleID, false);
+      this._shouldUnmockTransitiveDependenciesCache.set(key, true);
       return false;
     }
-
-    return (this._shouldMockModuleCache[moduleID] = true);
+    this._shouldMockModuleCache.set(moduleID, true);
+    return true;
   }
 
   private _createRequireImplementation(
     from: InitialModule,
     options?: InternalModuleOptions,
-  ): LocalModuleRequire {
+  ): NodeRequire {
     // TODO: somehow avoid having to type the arguments - they should come from `NodeRequire/LocalModuleRequire.resolve`
     const resolve = (moduleName: string, options: ResolveOptions) =>
       this._requireResolve(from.filename, moduleName, options);
     resolve.paths = (moduleName: string) =>
       this._requireResolvePaths(from.filename, moduleName);
 
-    const moduleRequire = (options && options.isInternalModule
+    const moduleRequire = (options?.isInternalModule
       ? (moduleName: string) =>
           this.requireInternalModule(from.filename, moduleName)
-      : this.requireModuleOrMock.bind(
-          this,
-          from.filename,
-        )) as LocalModuleRequire;
+      : this.requireModuleOrMock.bind(this, from.filename)) as NodeRequire;
     moduleRequire.extensions = Object.create(null);
-    moduleRequire.requireActual = this.requireActual.bind(this, from.filename);
-    moduleRequire.requireMock = this.requireMock.bind(this, from.filename);
     moduleRequire.resolve = resolve;
     moduleRequire.cache = (() => {
-      const notPermittedMethod = () => {
-        if (!this._hasWarnedAboutRequireCacheModification) {
-          this._environment.global.console.warn(
-            '`require.cache` modification is not permitted',
-          );
-
-          this._hasWarnedAboutRequireCacheModification = true;
-        }
-        return true;
-      };
-      return new Proxy<RequireCache>(Object.create(null), {
+      // TODO: consider warning somehow that this does nothing. We should support deletions, anyways
+      const notPermittedMethod = () => true;
+      return new Proxy<typeof moduleRequire['cache']>(Object.create(null), {
         defineProperty: notPermittedMethod,
         deleteProperty: notPermittedMethod,
         get: (_target, key) =>
@@ -1342,21 +1360,21 @@ class Runtime {
     };
     const unmock = (moduleName: string) => {
       const moduleID = this._resolver.getModuleID(
-        this._virtualMocks,
+        fromEntries(this._virtualMocks),
         from,
         moduleName,
       );
-      this._explicitShouldMock[moduleID] = false;
+      this._explicitShouldMock.set(moduleID, false);
       return jestObject;
     };
     const deepUnmock = (moduleName: string) => {
       const moduleID = this._resolver.getModuleID(
-        this._virtualMocks,
+        fromEntries(this._virtualMocks),
         from,
         moduleName,
       );
-      this._explicitShouldMock[moduleID] = false;
-      this._transitiveShouldMock[moduleID] = false;
+      this._explicitShouldMock.set(moduleID, false);
+      this._transitiveShouldMock.set(moduleID, false);
       return jestObject;
     };
     const mock: Jest['mock'] = (moduleName, mockFactory, options) => {
@@ -1365,11 +1383,11 @@ class Runtime {
       }
 
       const moduleID = this._resolver.getModuleID(
-        this._virtualMocks,
+        fromEntries(this._virtualMocks),
         from,
         moduleName,
       );
-      this._explicitShouldMock[moduleID] = true;
+      this._explicitShouldMock.set(moduleID, true);
       return jestObject;
     };
     const setMockFactory = (
@@ -1605,6 +1623,18 @@ class Runtime {
       xit: this._environment.global.xit,
       xtest: this._environment.global.xtest,
     };
+  }
+
+  private readFile(filename: Config.Path): string {
+    let source = this._cacheFS.get(filename);
+
+    if (!source) {
+      source = fs.readFileSync(filename, 'utf8');
+
+      this._cacheFS.set(filename, source);
+    }
+
+    return source;
   }
 }